// src/store/actions/types/usersTypes.ts
export enum ACTIONS_TYPE {
  FOLLOW_SUCCESS = 'Users/FOLLOW_SUCCESS',
  UNFOLLOW_SUCCESS = 'Users/UNFOLLOW_SUCCESS',
  SET_USERS = 'Users/SET_USERS',
  SET_CURRENT_PAGE = 'Users/SET_CURRENT_PAGE',
  SET_TOTAL_MEMBERS = 'Users/SET_TOTAL_MEMBERS',
  TOGGLE_IS_FETCHING = 'Users/TOGGLE_IS_FETCHING',
  TOGGLE_PRESSING_IN_PROGRESS = 'Users/TOGGLE_PRESSING_IN_PROGRESS',
  SET_USERS_FILTER = 'Users/SET_USERS_FILTER',
  SET_FOLLOWINGS = 'Users/SET_FOLLOWINGS',
  SET_FOLLOWINGS_COUNT = 'Users/SET_FOLLOWINGS_COUNT',
}